MERNIS Authentication Has Been Added.
- Added the MERNIS Authenticaton Service. MERNIS Authentication Service is a Web Service which checks the identity numbers of Turkish People.

- Some file names has been changed.

- A few if conditions has been added for more stable checking for the simulation.
For example we wanted to add new person to the database and that person is not authenticated. So the program would not add that person to database but as simulation we can edit that person's preferences. This makes a problem because the program did not add the person to database. Program has not that problem anymore.